common write module
